src/requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
attrs==24.2.0
bitsandbytes==0.44.1
blinker==1.9.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
datasets==3.1.0
date-spacy==0.0.1
dateparser==1.2.0
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
evaluate==0.4.3
faiss-cpu==1.9.0
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
hf_transfer==0.1.8
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.23.5
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langcodes==3.4.1
langsmith==0.1.143
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.13.5
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
orjson==3.10.11
packaging==24.2
pandas==2.2.2
peft==0.13.2
pillow==11.0.0
pip==24.0
pip3-autoremove==1.2.2
platformdirs==4.3.6
preshed==3.0.9
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
sentencepiece==0.2.0
sentry-sdk==2.18.0
seqeval==1.2.2
setproctitle==1.3.3
setuptools==65.5.1
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
span-marker==1.5.0
SQLAlchemy==2.0.35
srsly==2.4.8
starlette==0.41.2
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
triton==3.0.0
trl==0.12.1
typer==0.13.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tyro==0.8.14
tzdata==2024.2
tzlocal==5.2
unsloth==2024.11.7
unsloth_zoo==2024.11.5
urllib3==2.2.3
wandb==0.18.7
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wheel==0.45.0
wrapt==1.16.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.17.1